+++ NEWS.md
@@ -1,4 +1,5 @@
 ## RBedrock Development
+ * Added support for reading and writing 1.18.30+ actor digests.
  * Chunk data functions have been expanded.
  * [Breaking Change] Biome functions now return CNC biomes data (CnC biome functions have been renamed.) Accessing legacy biome data can be done with legacy biome functions.
  * [Breaking Change] Chunk data functions and files have been renamed to match internal Mojang names.